requirements.txt
# PS2 Texture Sorter - Dependencies
# Author: Dead On The Inside / JosephsDeadish

# UI Framework
customtkinter>=5.2.0
pillow>=10.0.0
darkdetect>=0.8.0

# Image Processing
opencv-python>=4.8.0
numpy>=1.24.0
scikit-image>=0.21.0

# DDS Support
Pillow>=10.0.0  # Has basic DDS support

# Machine Learning (lightweight)
scikit-learn>=1.3.0

# NEW: AI/ML - Offline model support
onnxruntime>=1.16.0  # Offline AI model inference

# Vision Models and Deep Learning
torch>=2.0.0  # PyTorch for deep learning models
torchvision>=0.15.0  # Computer vision models and transforms
transformers>=4.30.0  # CLIP, ViT, and other transformers
timm>=0.9.0  # EfficientNet, ResNet, and other vision models
open-clip-torch>=2.20.0  # Open source CLIP implementation

# Similarity Search
faiss-cpu>=1.7.4  # Vector similarity search (use faiss-gpu for CUDA)
chromadb>=0.4.0  # Vector database alternative
annoy>=1.17.0  # Approximate nearest neighbors

# Upscaling and Preprocessing
basicsr>=1.4.2  # Real-ESRGAN and super-resolution models
realesrgan>=0.3.0  # Real-ESRGAN upscaling

# File Operations
send2trash>=1.8.2
watchdog>=3.0.0

# Performance
psutil>=5.9.5
tqdm>=4.66.0
xxhash>=3.4.0  # Fast hashing for performance

# NEW: Archive Support
py7zr>=0.20.1  # 7z file support (>= 0.20.1 for security fix)

# NEW: Hotkeys
pynput>=1.7.6  # Global hotkeys

# NEW: Optional online AI
requests>=2.31.0  # For online API calls

# Utilities
colorama>=0.4.6
pyyaml>=6.0

# Build & Packaging
pyinstaller>=5.13.0
